+++ src/AutomaticOptionsProviderInterface.php
@@ -2,16 +2,20 @@
 namespace Consolidation\AnnotatedCommand;
 
 use Consolidation\AnnotatedCommand\Parser\CommandInfo;
+use Symfony\Component\Console\Input\InputOption;
 
 /**
- * Command cration listeners can be added to the annotation
- * command factory.  These will be notified whenever a new
- * commandfile is provided to the factory.  This is useful for
- * initializing new commandfile objects.
+ * Option providers can add options to commands based on the annotations
+ * present in a command.  For example, a command that specifies @fields
+ * will automatically be given --format and --fields options.
  *
  * @see AnnotatedCommandFactory::addListener()
+ * @see HookManager::addOptionHook()
  */
 interface AutomaticOptionsProviderInterface
 {
+    /**
+     * @return InputOption
+     */
     public function automaticOptions(CommandInfo $commandInfo);
 }